# docs/swagger.yaml
basePath: /
definitions:
...
  github_com_skiix-backend_internal_domain.Comment:
    properties:
      content:
        type: string
      created_at:
        type: string
      id:
        type: string
      post_id:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
  github_com_skiix-backend_internal_domain.CreatePostRequest:
    properties:
      categories:
        items:
          type: string
        type: array
      circle_id:
        type: string
      content:
        type: string
      image_url:
        type: string
      visibility:
        enum:
        - public
        - circle
        type: string
    required:
    - content
    - visibility
    type: object
...
  github_com_skiix-backend_internal_domain.Post:
    properties:
      circle_id:
        example: 550e8400-e29b-...
        type: string
      comment_count:
        example: 2
        type: integer
      content:
        example: Hello world
        type: string
      created_at:
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      image_url:
        example: https://cdn.skiix.com/image.jpg
        type: string
      like_count:
        example: 5
        type: integer
      updated_at:
        type: string
      user_id:
        example: 550e8400-e29b-41d4-a716-446655440001
        type: string
      visibility:
        description: public, circle
        example: public
        type: string
    type: object
  github_com_skiix-backend_internal_domain.PostFeedItem:
    properties:
      author_email:
        type: string
      circle_id:
        example: 550e8400-e29b-...
        type: string
      comment_count:
        example: 2
        type: integer
      content:
        example: Hello world
        type: string
      created_at:
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      image_url:
        example: https://cdn.skiix.com/image.jpg
        type: string
      is_liked:
        type: boolean
      like_count:
        example: 5
        type: integer
      updated_at:
        type: string
      user_id:
        example: 550e8400-e29b-41d4-a716-446655440001
        type: string
      visibility:
        description: public, circle
        example: public
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Skiix authentication API with JWT and OAuth2 support
  title: Skiix API
  version: "1.0"
paths:
  /api/v1/circles/{id}/posts:
    get:
      description: Get paginated feed of posts within a specific circle
      parameters:
      - description: Circle ID
        in: path
        name: id
        required: true
        type: string
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_skiix-backend_internal_domain.PostFeedItem'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get circle feed
      tags:
      - posts
      - circles
  /api/v1/feed:
    get:
      description: Get paginated feed of all posts
      parameters:
      - description: Number of posts to return (max 50, default 20)
        in: query
        name: limit
        type: integer
      - description: Number of posts to skip
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_skiix-backend_internal_domain.PostFeedItem'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get feed
      tags:
      - posts
...
  /api/v1/posts:
    post:
      consumes:
      - application/json
      description: Create a new post with optional image and categories
      parameters:
      - description: Post content
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_skiix-backend_internal_domain.CreatePostRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_skiix-backend_internal_domain.Post'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Create a new post
      tags:
      - posts
  /api/v1/posts/{post_id}:
    delete:
      description: Delete a post. Only the post owner can delete.
      parameters:
      - description: Post ID
        in: path
        name: post_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Delete a post
      tags:
      - posts
    get:
      description: Get a post by its ID
      parameters:
      - description: Post ID
        in: path
        name: post_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_skiix-backend_internal_domain.Post'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get a single post
      tags:
      - posts
    put:
      consumes:
      - application/json
      description: Update post content. Only the post owner can update.
      parameters:
      - description: Post ID
        in: path
        name: post_id
        required: true
        type: string
      - description: Updated content
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_skiix-backend_internal_domain.CreatePostRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_skiix-backend_internal_domain.Post'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Update a post
      tags:
      - posts
  /api/v1/posts/{post_id}/comments:
    get:
      description: Get paginated comments for a specific post
      parameters:
      - description: Post ID
        in: path
        name: post_id
        required: true
        type: string
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_skiix-backend_internal_domain.Comment'
            type: array
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get comments on a post
      tags:
      - comments
    post:
      consumes:
      - application/json
      description: Create a comment on a specific post
      parameters:
      - description: Post ID
        in: path
        name: post_id
        required: true
        type: string
      - description: Comment content
        in: body
        name: request
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_skiix-backend_internal_domain.Comment'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Add a comment to a post
      tags:
      - comments
  /api/v1/posts/{post_id}/comments/{comment_id}:
    delete:
      description: Delete a comment. Only the comment owner can delete.
      parameters:
      - description: Post ID
        in: path
        name: post_id
        required: true
        type: string
      - description: Comment ID
        in: path
        name: comment_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Delete a comment
      tags:
      - comments
  /api/v1/posts/{post_id}/like:
    post:
      description: Like a post if not liked, unlike if already liked
      parameters:
      - description: Post ID
        in: path
        name: post_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Toggle like on a post
      tags:
      - posts
  /api/v1/users/{user_id}/posts:
    get:
      description: Get all posts by a specific user
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_skiix-backend_internal_domain.PostFeedItem'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get posts by user
      tags:
      - posts
...
schemes:
- http
- https
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token
    in: header
    name: Authorization
    type: apiKey
    x-tokenname: Authorization
swagger: "2.0"